Refactor secret form and homepage structure
Extract core secret creation form logic into a reusable `SecretForm.vue` component. This replaces `BaseSecretFormLite.vue`. Update the homepage form component (`HomepageSecretForm.vue`, previously `SecretFormLite.vue`) to wrap the new generic form. Introduce `SecretRegionExperience.vue` to encapsulate the region selector and the form on the homepage, improving component composition.
