+++ src/components/vue/homepage/Homepage.vue
@@ -13,9 +13,9 @@ import { setLanguage } from "@/i18n";
 import { setLanguageWithMessages, type MessageSchema } from "@/i18n";
 import { useJurisdiction } from "@/composables/useJurisdiction";
 
-import SecretFormLite from "@/components/vue/homepage/SecretFormLite.vue";
+import SecretRegionExperience from "@/components/vue/homepage/SecretRegionExperience.vue";
 import UseCaseSelector from "@/components/vue/homepage/UseCaseSelector.vue";
-import type { ApiResult } from "@/components/vue/forms/BaseSecretFormLite.vue";
+import type { ApiResult } from "@/components/vue/forms/SecretForm.vue";
 
 const props = defineProps<{
   locale: string;
@@ -65,6 +65,13 @@ const switchRegion = (newRegion?: string | { identifier: string }) => {
   }
 };
 
+// Handle region change from the SecretRegionExperience component
+const handleRegionChange = (region: any) => {
+  if (region && region.identifier) {
+    setJurisdiction(region.identifier);
+  }
+};
+
 const secretFormRef = ref();
 
 const handleSecretCreationResult = (result: ApiResult) => {
@@ -107,53 +114,16 @@ onUnmounted(() => {
       <!-- Section 1: Branding and Benefits -->
       <HeroTitle />
 
-      <!-- Section 2: Secret Form Lite with Region Selector -->
-      <section class="bg-gradient-to-b from-gray-50 to-white dark:from-gray-800 dark:to-gray-900 py-8">
-        <div class="mx-auto max-w-5xl px-4 sm:px-6 lg:px-8">
-          <div
-            class="bg-white dark:bg-gray-800 shadow-xl rounded-xl overflow-hidden border border-gray-100 dark:border-gray-700">
-            <div
-              class="bg-gradient-to-r from-brand-500/10 to-brand-600/5 dark:from-brand-500/20 dark:to-brand-600/10 px-6 py-4 border-b border-gray-200 dark:border-gray-700">
-              <div class="flex flex-wrap justify-between items-center gap-4">
-                <h3 class="text-lg font-semibold text-gray-900 dark:text-white">
-                  {{
-                    t("LABELS.create-link") ||
-                    "Create a secure, self-destructing message"
-                  }}
-                </h3>
-
-                <!-- Premium region selector with enhanced styling and animations -->
-                <div class="flex-shrink-0">
-                  <ClientOnlyRegionSelector
-                    v-if="isClient"
-                    :current-region="currentRegion"
-                    :available-regions="availableRegions"
-                    class="rounded-lg px-2 py-1.5 bg-white/90 dark:bg-gray-700/90 border border-gray-200 dark:border-gray-600 shadow-sm"
-                    @region-change="region => switchRegion(region.identifier)" />
-                </div>
-              </div>
-            </div>
-
-            <div class="px-6 py-5">
-              <SecretFormLite
-                ref="secretFormRef"
-                :placeholder="
-                  t('web.secrets.secretPlaceholder-premium', {
-                    noun: currentRegion.displayName,
-                  })
-                "
-                :api-base-url="apiBaseUrl"
-                :with-options="false"
-                @create-link="handleSecretCreationResult" />
-            </div>
-
-            <div
-              class="bg-gray-50 dark:bg-gray-700 px-6 py-3 text-xs text-center text-gray-500 dark:text-gray-400 border-t border-gray-100 dark:border-gray-600">
-              {{ t("web.secrets.complianceNote") }}
-            </div>
-          </div>
-        </div>
-      </section>
+      <!-- Section 2: Secret Form with Region Selector -->
+      <SecretRegionExperience
+        ref="secretFormRef"
+        :current-region="currentRegion"
+        :available-regions="availableRegions"
+        :api-base-url="apiBaseUrl"
+        :is-client="isClient"
+        @region-change="handleRegionChange"
+        @create-secret="handleSecretCreationResult"
+      />
 
       <!-- Section 3: Feature Highlights -->
       <FeatureHighlights />
